refactor: centralize evaluation identity
